# openapi/admin.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ultrafast Gateway Admin API",
    "description": "The admin API of the Ultrafast gateway, served under /api.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "2.0.0-alpha.1"
  },
  "paths": {
...
    "/api/tokens": {
      "get": {
        "tags": [
          "tokens"
        ],
        "operationId": "tokens_list",
        "responses": {
          "200": {
            "description": "The caller's access tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenList"
                }
              }
            }
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      },
      "post": {
        "tags": [
          "tokens"
        ],
        "operationId": "tokens_create",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "description": "The CSRF token of the session. Required with a session cookie; not needed with an access token.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The new access token, with the token itself.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedToken"
                }
              }
            }
          },
          "400": {
            "description": "The request is not of the expected form.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not allowed to do this, or the CSRF token is missing or does not match.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The request body is too large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are not valid; `fields` names each of them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      }
    },
    "/api/tokens/{id}": {
      "delete": {
        "tags": [
          "tokens"
        ],
        "operationId": "tokens_revoke",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the access token.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "description": "The CSRF token of the session. Required with a session cookie; not needed with an access token.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The access token is revoked."
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not allowed to do this, or the CSRF token is missing or does not match.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "It does not exist, or it is hidden from the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Every error of `/api` has this shape.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable name for the error, such as `not_found`."
          },
          "fields": {
            "type": "object",
            "description": "For `validation_failed` only: a message for each field that is not\nvalid.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "message": {
            "type": "string",
            "description": "Text for a person. It may change."
          }
        }
      },
...
      "CreateTokenRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CreatedToken": {
        "type": "object",
        "required": [
          "token",
          "secret"
        ],
        "properties": {
          "secret": {
            "type": "string",
            "description": "The access token itself. It is shown once, in this answer, and\ncannot be read again."
          },
          "token": {
            "$ref": "#/components/schemas/TokenView"
          }
        }
      },
...
      "TokenList": {
        "type": "object",
        "required": [
          "tokens"
        ],
        "properties": {
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenView"
            }
          }
        }
      },
      "TokenStatus": {
        "type": "string",
        "description": "Whether a token can still be used, as of the moment of the answer.",
        "enum": [
          "active",
          "expired",
          "revoked"
        ]
      },
      "TokenView": {
        "type": "object",
        "description": "A token as `/api` shows it. It has no field for the token or its hash.",
        "required": [
          "id",
          "name",
          "display",
          "expires_at",
          "revoked_at",
          "last_used_at",
          "created_at",
          "status"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "display": {
            "type": "string"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TokenStatus",
            "description": "Worked out when the answer is made."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uf_session",
        "description": "The session cookie set by signing in."
      },
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "An access token, which starts with `uf-at-`."
      }
    }
  },
...
}